requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-pandas==0.6.6
djangorestframework==3.14.0
numpy==1.24.2
odfpy==1.4.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.2
six==1.16.0
sqlparse==0.4.3
